Simplifies code of Parameters, CommandlineOptions, Tests
Declaration of parameters sets was too complicated, three new
identifiers needed to be created. Now it's just a matter of copying a
default, modify as needed and append to a list.

CommandlineOptions dropped suport to long option names, as it was not
useful. Created a new option to list available parameterizations.

Created a new file to hold all testing.

This revision won't compile.