Fixes reproduction of result for G_V != 0 from Buballa's report
For some reason, changing the sign of the contribution proportional to $G_V$ in the thermodynamic potential makes the result nearly identical with those in the reference. Also, if the chemical potential is zero, then the renormalized chemical potential is also zero, as this is the only possible solution for the self-consistent equation on the renormalized chemical potential. Fixed the special case treatment.