Adds implementation of finite temperature functions
Adds functions necessary for the calculation of the thermodynamic potential and of the equations of state. Integrates de finite temperature in a very simple way. Maybe it would be better to implement the zero temperature version of each relevant function as a special case of the finite temperature ones.

Needs testings, verification of dimensions.