Fix calculation of entropy
Reverted to old method of calculation of the integrand of the entropy density, as it appears to be more stable. There was a sign that was incorrect, but once fixed it gave the same results as both other methods, but with better numerical stability at low momentum for some cases.