Save last root and use it as a guess for the next
In the finite temperature case, when we determine a root for the
simultaneous solving of the mass and density gap equations, save
the obtained root and use it as a guess for the next iteration.
As we step the variable sequentially, it means that the next root
is close to the one just calculated. Using the present root as a
guess for the next greatly speed up the execution and minimizes the
chance of having a "solver stuck" problem.