Generate new data
This is mostly to compare with Tsue arXiv:1003.2843v1. Even though we use different parameterizations, the results are quite similar. The reproduction of Fig. 12 (not included in the graphics here, but data/pressure.dat is part of the data necessary for the reproduction) is pretty dissimilar. Maybe it's just the parameterizations used.