Fix sign to be like Buballa
There was an error (a chemical potential instead of a renormalized
chemical potential) in the expression for the thermodynamic potential
at zero temperature. This was corrected by trial and error by changing
the sign of the term on $G_V$. This solution turns out to exactly
compensate the exchange of the chemical potentials.

However, for clarity, it's better to stick to the reference. Another point
is that I changed the corresponding sign for the finite temperature also.
This was an error.

Also fixes the plotting scripts to not show unnecessary keys, as some _
characters in it were generating problems in latex (tikz versions of the
plots)